package/iozone: bump to version 3.508
iozone releases 3.507 and 3.508 seems to be only build fixes releases
for latest GCC compiler version but since there is no public vcs
it's not easy to review the history between releases and backport
any patches.

Based on the changelog from [1]:

  Revision 3.507
  Fix GCC compile warnings.

  Revision 3.508
  Put an end to the (&*% stupid GCC breaking builds for no valid reason.

So bump to the latest 508 release.

Rebase 0001-Add-new-targets-for-iozone.patch
Rebase 0002-fix-build-without-aio.patch and convert to git format

The TestIozone build issue is not yet fixed by the version bump [2].

[1] https://www.iozone.org/src/current/Changes.txt
[2] https://gitlab.com/buildroot.org/buildroot/-/jobs/11176774405

Signed-off-by: Romain Naour <romain.naour@smile.fr>
Signed-off-by: Julien Olivain <ju.o@free.fr>